+++ gazpacho/download.py
@@ -6,6 +6,7 @@ from urllib.request import urlopen
 
 # does this make sense? for multiple usage, try with wav? and videos?
 
+
 class Bytes(BytesIO):
     def __init__(self, url):
         contents = urlopen(url).read()
@@ -21,6 +22,7 @@ def download(url, path=None):
     if not path:
         return Bytes(url)
     urlretrieve(url, path)
+
 
 # from gazpacho import download
 # asset()